Add files via upload
The goal of this project is to analyse the given Skin Cancer Image Data and predict the diseases. The data was divided into two parts i.e. the training and the test data set. Models were developed based on the training data set and applied to the test data set to find out the accuracy of each model based on the predicted values generated.  Based on these values we can determine how good a particular model is for prediction of Skin Cancer